+++ Ryujinx.Audio/Renderer/Dsp/Command/PerformanceCommand.cs
@@ -17,7 +17,7 @@ namespace Ryujinx.Audio.Renderer.Dsp.Command
 
         public CommandType CommandType => CommandType.Performance;
 
-        public ulong EstimatedProcessingTime { get; set; }
+        public uint EstimatedProcessingTime { get; set; }
 
         public PerformanceEntryAddresses PerformanceEntryAddresses { get; }
