+++ Ryujinx.Audio/Renderer/Dsp/Command/ReverbCommand.cs
@@ -34,7 +34,7 @@ namespace Ryujinx.Audio.Renderer.Dsp.Command
 
         public CommandType CommandType => CommandType.Reverb;
 
-        public ulong EstimatedProcessingTime { get; set; }
+        public uint EstimatedProcessingTime { get; set; }
 
         public ReverbParameter Parameter => _parameter;
         public Memory<ReverbState> State { get; }
